implemented interface for taxes

+++ src/personen/Mensch.java
@@ -13,7 +13,7 @@ public abstract class Mensch extends Einwohner {
 	protected String jobs;
 	protected int alter;
 
-	public Mensch(String nachname, String vorname, double einkommen, String jobs, int alter) throws NegativeIncomeException {
+	public Mensch(String nachname, String vorname, int einkommen, String jobs, int alter) throws NegativeIncomeException {
 		super(nachname, vorname, einkommen);
 		this.jobs = jobs;
 		this.alter = alter;
